Refactored library name - Fixed bug that prevented test running
