+++ f1tenth_gym/envs/reset/__init__.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
-from f110_gym.envs.reset.masked_reset import GridResetFn, AllTrackResetFn
-from f110_gym.envs.reset.reset_fn import ResetFn
-from f110_gym.envs.track import Track
+from .masked_reset import GridResetFn, AllTrackResetFn
+from .reset_fn import ResetFn
+from ..track import Track
 
 
 def make_reset_fn(type: str | None, track: Track, num_agents: int, **kwargs) -> ResetFn:
