+++ tests/test_collision_checks.py
@@ -3,7 +3,7 @@ import time
 import unittest
 
 import numpy as np
-from f110_gym.envs.collision_models import get_vertices, collision
+from f1tenth_gym.envs.collision_models import get_vertices, collision
 
 
 # Copyright (c) 2020 Joseph Auckley, Matthew O'Kelly, Aman Sinha, Hongrui Zheng
